move to a munki package
